+++ data/lexicon/switches/noun-verb.js
@@ -818,6 +818,7 @@ export default [
   'wipe',
   'wire',
   'wish',
+  'wound',
   'wonder',
   'work',
   'worry',