+++ src/2-two/preTagger/methods/transform/index.js
@@ -6,5 +6,4 @@ export default {
   noun: nouns,
   verb: verbs,
   adjective: adjectives
-
 }