+++ src/API/extend.js
@@ -37,43 +37,43 @@ const addIrregulars = function (model, conj) {
     // verb forms
     if (conj[k].pastTense) {
       if (m.toPast) {
-        m.toPast.exceptions[k] = conj[k].pastTense
+        m.toPast.ex[k] = conj[k].pastTense
       }
       if (m.fromPast) {
-        m.fromPast.exceptions[conj[k].pastTense] = k
+        m.fromPast.ex[conj[k].pastTense] = k
       }
     }
     if (conj[k].presentTense) {
       if (m.toPresent) {
-        m.toPresent.exceptions[k] = conj[k].presentTense
+        m.toPresent.ex[k] = conj[k].presentTense
       }
       if (m.fromPresent) {
-        m.fromPresent.exceptions[conj[k].presentTense] = k
+        m.fromPresent.ex[conj[k].presentTense] = k
       }
     }
     if (conj[k].gerund) {
       if (m.toGerund) {
-        m.toGerund.exceptions[k] = conj[k].gerund
+        m.toGerund.ex[k] = conj[k].gerund
       }
       if (m.fromGerund) {
-        m.fromGerund.exceptions[conj[k].gerund] = k
+        m.fromGerund.ex[conj[k].gerund] = k
       }
     }
     // adjective forms
     if (conj[k].comparative) {
       if (m.toComparative) {
-        m.toComparative.exceptions[k] = conj[k].comparative
+        m.toComparative.ex[k] = conj[k].comparative
       }
       if (m.fromComparative) {
-        m.fromComparative.exceptions[conj[k].comparative] = k
+        m.fromComparative.ex[conj[k].comparative] = k
       }
     }
     if (conj[k].superlative) {
       if (m.toSuperlative) {
-        m.toSuperlative.exceptions[k] = conj[k].superlative
+        m.toSuperlative.ex[k] = conj[k].superlative
       }
       if (m.fromSuperlative) {
-        m.fromSuperlative.exceptions[conj[k].superlative] = k
+        m.fromSuperlative.ex[conj[k].superlative] = k
       }
     }
   })