+++ tests/three/api.test.js
@@ -65,7 +65,6 @@ test('api:', function (t) {
     doc.post(' ')
     doc.trim()
     doc.hyphenate()
-    doc.debullet()
     doc.dehyphenate()
     doc.toQuotations()
     doc.toParentheses()
@@ -189,6 +188,7 @@ test('api:', function (t) {
     doc.sentences().toFutureTense()
     doc.sentences().toInfinitive()
     doc.sentences().toNegative()
+    doc.sentences().debullet()
     doc.questions()
 
     // verbs