+++ changelog.md
@@ -18,11 +18,16 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 <!-- #### 14.8.1 [Unreleased]
 -->
 
-#### 14.9.0 [Mar 2023]
+#### 14.9.0 [May 2023]
+- **[new]** - .verbs().toPastParticiple() method
+- **[new]** - .debullet() method
 - **[change]** - typescript path changes (thanks @rotemdan !)
+- **[fix]** - suffix tagging issues
+- **[fix]** - match syntax #997
+- **[change]** - keep possessive in replace #1011
+- **[update]** - dependencies
 - **[fix]** - parsematch bug #997
 - **[fix]** - tagger fix #998
-- **[new]** - .debullet() method
 
 #### 14.8.2 [Feb 2023]
 - **[change]** - #Actor tagging - in advance of #565